refactor(): move types to folder
