+++ src/features/article/components/atoms/Breadcrumbs.tsx
@@ -1,6 +1,4 @@
-type BreadcrumbsType = {
-    link: string;
-};
+import { BreadcrumbsType } from '../../types/article.types';
 
 export default function Breadcrumbs({ link }: BreadcrumbsType) {
     return <p>{link}</p>;
